Cleanup configure and Makefile.in

+++ gzread.c
@@ -286,7 +286,7 @@ static size_t gz_read(gz_statep state, void *buf, size_t len) {
         /* set n to the maximum amount of len that fits in an unsigned int */
         n = -1;
         if (n > len)
-            n = len;
+            n = (unsigned)len;
 
         /* first just try copying data from the output buffer */
         if (state->x.have) {
@@ -363,7 +363,7 @@ int ZEXPORT gzread(gzFile file, void *buf, unsigned len) {
     }
 
     /* read len or fewer bytes to buf */
-    len = gz_read(state, buf, len);
+    len = (unsigned)gz_read(state, buf, len);
 
     /* check for an error */
     if (len == 0 && state->err != Z_OK && state->err != Z_BUF_ERROR)
@@ -423,7 +423,7 @@ int ZEXPORT gzgetc(gzFile file) {
     }
 
     /* nothing there -- try gz_read() */
-    ret = gz_read(state, buf, 1);
+    ret = (int)gz_read(state, buf, 1);
     return ret < 1 ? -1 : buf[0];
 }
